+++ src/main/java/pro/trevor/tankgame/rule/impl/ruleset/DefaultV3RulesetRegister.java
@@ -1,4 +1,4 @@
-package pro.trevor.tankgame.rule.impl.version4;
+package pro.trevor.tankgame.rule.impl.ruleset;
 
 import pro.trevor.tankgame.rule.definition.ApplicableRuleset;
 import pro.trevor.tankgame.rule.definition.MetaTickActionRule;
@@ -7,12 +7,9 @@ import pro.trevor.tankgame.rule.definition.enforcer.EnforcerRuleset;
 import pro.trevor.tankgame.rule.definition.enforcer.MaximumEnforcer;
 import pro.trevor.tankgame.rule.definition.enforcer.MinimumEnforcer;
 import pro.trevor.tankgame.rule.definition.player.PlayerRuleset;
-import pro.trevor.tankgame.rule.definition.player.TimedPlayerActionRule;
-import pro.trevor.tankgame.rule.impl.IRulesetRegister;
-import pro.trevor.tankgame.rule.impl.shared.rule.ConditionalRules;
-import pro.trevor.tankgame.rule.impl.shared.rule.PlayerRules;
-import pro.trevor.tankgame.rule.impl.shared.rule.TickRules;
-import pro.trevor.tankgame.rule.impl.util.BaseRulesetRegister;
+import pro.trevor.tankgame.rule.impl.shared.ConditionalRules;
+import pro.trevor.tankgame.rule.impl.shared.PlayerRules;
+import pro.trevor.tankgame.rule.impl.shared.TickRules;
 import pro.trevor.tankgame.state.State;
 import pro.trevor.tankgame.state.attribute.Attribute;
 import pro.trevor.tankgame.state.board.Board;
@@ -22,12 +19,10 @@ import pro.trevor.tankgame.state.meta.Council;
 import pro.trevor.tankgame.state.meta.PlayerRef;
 import pro.trevor.tankgame.util.RulesetType;
 
-import java.util.function.Function;
+import static pro.trevor.tankgame.rule.impl.shared.TickRules.INCREMENT_DAY_ON_TICK;
 
-@RulesetType(name = "default-v4")
-public class RulesetRegister extends BaseRulesetRegister implements IRulesetRegister {
-
-    private static final Function<State, Long> TIMEOUT = (s) -> (long) (5 * 60);
+@RulesetType(name = "default-v3")
+public class DefaultV3RulesetRegister extends BaseRulesetRegister implements IRulesetRegister {
 
     @Override
     public void registerEnforcerRules(Ruleset ruleset) {
@@ -41,6 +36,7 @@ public class RulesetRegister extends BaseRulesetRegister implements IRulesetRegi
         invariants.put(GenericTank.class, new MaximumEnforcer<>(Attribute.ACTION_POINTS, 5));
         invariants.put(GenericTank.class, new MinimumEnforcer<>(Attribute.BOUNTY, 0));
         invariants.put(BasicWall.class, new MinimumEnforcer<>(Attribute.DURABILITY, 0));
+
         invariants.put(Council.class, new MinimumEnforcer<>(Attribute.COFFER, 0));
     }
 
@@ -50,27 +46,9 @@ public class RulesetRegister extends BaseRulesetRegister implements IRulesetRegi
 
         tickRules.put(GenericTank.class, TickRules.GetDistributeGoldToTanksRule());
         tickRules.put(GenericTank.class, TickRules.GetGrantActionPointsOnTickRule(1));
-
-        tickRules.put(Board.class, TickRules.INCREMENT_DAY_ON_TICK);
         tickRules.put(Board.class, TickRules.GOLD_MINE_REMAINDER_GOES_TO_COFFER);
-        tickRules.put(Council.class, TickRules.GetCouncilBaseIncomeRule(1, 3));
-        tickRules.put(Council.class, TickRules.ARMISTICE_VIA_COUNCIL);
+        tickRules.put(Board.class, INCREMENT_DAY_ON_TICK);
         tickRules.put(Council.class, new MetaTickActionRule<>((s, c) -> Attribute.CAN_BOUNTY.to(c, true)));
-    }
-
-    @Override
-    public void registerPlayerRules(Ruleset ruleset) {
-        PlayerRuleset playerRules = ruleset.getPlayerRules();
-
-        playerRules.put(GenericTank.class, new TimedPlayerActionRule<>(PlayerRules.SHOOT_V4, TIMEOUT));
-        playerRules.put(GenericTank.class, new TimedPlayerActionRule<>(PlayerRules.GetMoveRule(Attribute.ACTION_POINTS, 1), TIMEOUT));
-        playerRules.put(GenericTank.class, new TimedPlayerActionRule<>(PlayerRules.GetShareGoldWithTaxRule(1), TIMEOUT));
-        playerRules.put(GenericTank.class, new TimedPlayerActionRule<>(PlayerRules.BuyActionWithGold(3, 1), TIMEOUT));
-        playerRules.put(GenericTank.class, new TimedPlayerActionRule<>(PlayerRules.GetUpgradeRangeRule(Attribute.GOLD, 5), TIMEOUT));
-
-        playerRules.put(Council.class, PlayerRules.GetCofferCostStimulusRule(3));
-        playerRules.put(Council.class, PlayerRules.GetRuleCofferCostGrantLife(15, 3));
-        playerRules.put(Council.class, PlayerRules.GetRuleCofferCostBounty(1, 5));
     }
 
     @Override
@@ -78,9 +56,21 @@ public class RulesetRegister extends BaseRulesetRegister implements IRulesetRegi
         ApplicableRuleset conditionalRules = ruleset.getConditionalRules();
         conditionalRules.put(GenericTank.class, ConditionalRules.GetKillOrDestroyTankOnZeroDurabilityRule());
         conditionalRules.put(BasicWall.class, ConditionalRules.DESTROY_WALL_ON_ZERO_DURABILITY);
-
-        conditionalRules.put(Council.class, ConditionalRules.ARMISTICE_COUNCIL_WIN_CONDITION);
         conditionalRules.put(Board.class, ConditionalRules.TANK_WIN_CONDITION);
+    }
+
+    @Override
+    public void registerPlayerRules(Ruleset ruleset) {
+        PlayerRuleset playerRules = ruleset.getPlayerRules();
+        playerRules.put(GenericTank.class, PlayerRules.BUY_ACTION_WITH_GOLD_PLUS_DISCOUNT);
+        playerRules.put(GenericTank.class, PlayerRules.GetUpgradeRangeRule(Attribute.GOLD, 8));
+        playerRules.put(GenericTank.class, PlayerRules.GetShareGoldWithTaxRule(1));
+        playerRules.put(GenericTank.class, PlayerRules.GetMoveRule(Attribute.ACTION_POINTS, 1));
+        playerRules.put(GenericTank.class, PlayerRules.SHOOT_V3);
+
+        playerRules.put(Council.class, PlayerRules.GetCofferCostStimulusRule(3));
+        playerRules.put(Council.class, PlayerRules.GetRuleCofferCostGrantLife(15, 3));
+        playerRules.put(Council.class, PlayerRules.GetRuleCofferCostBounty(1, 5));
     }
 
     @Override
