+++ src/test/java/pro/trevor/tankgame/rule/CouncilBountyTest.java
@@ -2,7 +2,7 @@ package pro.trevor.tankgame.rule;
 
 import org.junit.jupiter.api.Test;
 import pro.trevor.tankgame.rule.definition.player.PlayerActionRule;
-import pro.trevor.tankgame.rule.impl.shared.rule.PlayerRules;
+import pro.trevor.tankgame.rule.impl.shared.PlayerRules;
 import pro.trevor.tankgame.state.State;
 import pro.trevor.tankgame.state.attribute.Attribute;
 import pro.trevor.tankgame.state.meta.Council;
